+++ src/app/armory/LazyArmory.ts
@@ -1,3 +1,3 @@
 import { lazy } from 'react';
 
-export default lazy(() => import('./Armory' /* webpackChunkName: "armory" */));
+export default lazy(() => import('./Armory' /* webpackChunkName: "item-popup-armory" */));
