+++ src/app/dim-ui/WeaponGroupingIcon.tsx
@@ -1,5 +1,6 @@
 import TagIcon from 'app/inventory/TagIcon';
-import { AmmoIcon } from 'app/item-popup/ItemPopupHeader';
+import { AmmoIcon } from 'app/item-popup/AmmoIcon';
+import {} from 'app/item-popup/ItemPopupHeader';
 import { VaultGroupIcon } from 'app/shell/item-comparators';
 import ElementIcon from './ElementIcon';
 import { getWeaponTypeSvgIconFromCategoryHashes } from './svgs/itemCategory';
