+++ src/app/item-popup/AmmoIcon.m.scss
@@ -1,0 +1,11 @@
+@use '../variables.scss' as *;
+
+.ammoIcon {
+  margin: 0 2px 0 6px;
+  height: 14px;
+  width: 19px;
+}
+
+.primary {
+  opacity: 0.9;
+}
