+++ src/app/item-popup/AmmoIcon.m.scss.d.ts
@@ -1,0 +1,8 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'ammoIcon': string;
+  'primary': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
