+++ src/app/item-popup/ItemPopupHeader.m.scss
@@ -69,22 +69,12 @@
   }
 }
 
-.ammoIcon {
-  margin: 0 2px 0 6px;
-  height: 14px;
-  width: 19px;
-}
-
 .itemType {
   opacity: 0.7;
 }
 
 .power {
   margin: 0 4px 0 2px;
-}
-
-.primary {
-  opacity: 0.9;
 }
 
 .masterwork {
