+++ src/app/item-popup/ItemPopupHeader.m.scss.d.ts
@@ -1,7 +1,6 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'ammoIcon': string;
   'armory': string;
   'common': string;
   'details': string;
@@ -12,7 +11,6 @@ interface CssExports {
   'legendary': string;
   'masterwork': string;
   'power': string;
-  'primary': string;
   'pursuit': string;
   'rare': string;
   'subtitle': string;
