--- src/app/vendors/components.ts
@@ -1,5 +1,0 @@
-// A single file for the vendor components so we can load them as one chunk
-
-export { default as SingleVendorPage } from './single-vendor/SingleVendorPage';
-export { default as SingleVendorSheet } from './single-vendor/SingleVendorSheet';
-export { default as Vendors } from './Vendors';
